feat: modalStore 수정
