--- src/components/Modal/stores/modalStore.ts
@@ -1,32 +1,0 @@
-import { createStore } from 'zustand';
-import { useStoreWithEqualityFn as useStore } from 'zustand/traditional';
-
-const initialModalData = {
-  title: '',
-  firstButtonText: '',
-};
-
-interface ModalData {
-  title: string;
-  description?: string;
-  firstButtonText: string;
-  secondButtonText?: string;
-}
-
-interface State {
-  modalData: ModalData;
-}
-
-interface Action {
-  openModal: (modalData: ModalData) => void;
-}
-
-export const modalStore = createStore<State & Action>((set) => ({
-  modalData: initialModalData,
-  openModal: (modalData) => set({ modalData }),
-}));
-
-export const useModalStore = <T>(
-  selector: (state: State & Action) => T,
-  equals?: (a: T, b: T) => boolean,
-) => useStore(modalStore, selector, equals);
